pyproject.toml
[project]
name = "multi-tasker"
version = "0.1.0"
description = ""
authors = [
    {name = "redun",email = "redunsasha2017@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pyjwt[crypto] (>=2.11.0,<3.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "black (>=26.3.0,<27.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "redis (>=7.3.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
